src/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.6
imageio==2.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.16.4
opencv-python==4.1.2.30
openslide-python==1.1.1
pandas==0.25.3
Pillow==6.2.1
protobuf==3.8.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.5
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.2
scikit-image==0.16.2
scipy==1.3.3
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.40.2
Werkzeug==0.15.4
wrapt==1.11.2
